fix/windows: use `size_t` instead of `u64` where necessary
`size_t` depends on the target architecture.
Using `u64` causes the build to fail when targetting 32bit.

+++ src/frameworks/cuda/context.rs
@@ -72,7 +72,7 @@ impl IDevice for Context {
     }
 
     fn alloc_memory(&self, size: u64) -> Result<Memory, DeviceError> {
-        Ok(try!(API::mem_alloc(size)))
+        Ok(try!(API::mem_alloc(size as size_t)))
     }
 
     fn sync_in(&self, source: &DeviceType, source_data: &MemoryType, dest_data: &mut Memory) -> Result<(), DeviceError> {
